+++ src/helpers/getRandomCoordinate.tsx
@@ -1,0 +1,10 @@
+import { getRandomInt } from './getRandomInt';
+
+// получить рандомную координату из списка возможных
+export const getRandomCoordinate = (
+  availableCoordinatesArray: number[][]
+): number[] => {
+  return availableCoordinatesArray[
+    getRandomInt(availableCoordinatesArray.length)
+  ];
+};
